+++ src/LinqAn.Google/Metrics/Goal7Completions.cs
@@ -6,7 +6,7 @@ namespace LinqAn.Google.Metrics
 	/// 	The total number of completions for the requested goal number.
 	/// </summary>
 	[Description("The total number of completions for the requested goal number.")]
-	public class Goal7Completions: Metric<float>
+	public class Goal7Completions: Metric<int>
 	{
 		/// <summary>
 		/// 	Instantiates a <seealso cref="Goal7Completions" />.
